requirements.txt
openai>=1.0.0
jinja2>=3.0.0
pyyaml>=6.0
tqdm>=4.0
